+++ src/outputs/output_utils.hpp
@@ -19,19 +19,24 @@
 #define OUTPUTS_OUTPUT_UTILS_HPP_
 
 // C++
+#include <algorithm>
 #include <array>
+#include <functional>
 #include <map>
 #include <memory>
+#include <numeric>
 #include <set>
 #include <sstream>
 #include <string>
 #include <tuple>
+#include <utility>
 #include <vector>
 
 // Parthenon
 #include "basic_types.hpp"
 #include "interface/metadata.hpp"
 #include "interface/variable.hpp"
+#include "kokkos_abstraction.hpp"
 #include "mesh/domain.hpp"
 #include "mesh/mesh.hpp"
 #include "mesh/meshblock.hpp"
@@ -41,31 +46,93 @@ namespace parthenon {
 namespace OutputUtils {
 // Helper struct containing some information about a variable
 struct VarInfo {
+ public:
+  static constexpr int VNDIM = MAX_VARIABLE_DIMENSION;
   std::string label;
   int num_components;
-  int nx6;
-  int nx5;
-  int nx4;
-  int nx3;
-  int nx2;
-  int nx1;
   int tensor_rank; // 0- to 3-D for cell-centered variables, 0- to 6-D for arbitrary shape
                    // variables
   MetadataFlag where;
   bool is_sparse;
   bool is_vector;
+  IndexShape cellbounds;
   std::vector<std::string> component_labels;
-  int Size() const { return nx6 * nx5 * nx4 * nx3 * nx2 * nx1; }
+  // list of topological elements in variable... e.g., Face1, Face2, etc
+  std::vector<TopologicalElement> topological_elements;
+  // how many topological elements are stored in variable, e.g., 3 for
+  // face and edge vars.
+  int ntop_elems;
+  // whether or not topological element matters.
+  bool element_matters;
+
+  Triple_t<int> GetNumKJI(const IndexDomain domain) const;
+  Triple_t<IndexRange> GetPaddedBoundsKJI(const IndexDomain domain) const;
+
+  int Size() const;
+  // Includes topological element shape
+  int TensorSize() const;
+  // Size of region that needs to be filled with 0s if not allocated
+  int FillSize(const IndexDomain domain) const;
+  // number of elements of data that describe variable shape
+  int GetNDim() const;
+
+  template <typename T>
+  int FillShape(const IndexDomain domain, T *data) const {
+    if (where == MetadataFlag({Metadata::None})) {
+      for (int i = 0; i < tensor_rank; ++i) {
+        data[i] = static_cast<T>(rnx_[rnx_.size() - tensor_rank + i]);
+      }
+    } else {
+      // For nx1,nx2,nx3 find max storage required in each direction
+      // accross topological elements. Unused indices will be written but
+      // empty.
+      auto [nx3, nx2, nx1] = GetNumKJI(domain);
+      // fill topological element, if relevant
+      if (element_matters) {
+        data[0] = ntop_elems;
+      }
+      // fill the tensor rank
+      for (int i = 0; i < tensor_rank; ++i) {
+        data[i + element_matters] =
+            static_cast<T>(rnx_[rnx_.size() - 3 - tensor_rank + i]);
+      }
+      // fill cell indices
+      data[tensor_rank + element_matters] = static_cast<T>(nx3);
+      data[tensor_rank + element_matters + 1] = static_cast<T>(nx2);
+      data[tensor_rank + element_matters + 2] = static_cast<T>(nx1);
+    }
+    return GetNDim();
+  }
+
+  template <typename T, typename... Args>
+  int FillShape(const IndexDomain domain, T *head, Args... args) const {
+    int ndim_head = FillShape(domain, head);
+    int ndim_tail = FillShape(domain, std::forward<Args>(args)...);
+    // this check should be impossible to trigger but... just to be safe
+    PARTHENON_DEBUG_REQUIRE(ndim_head == ndim_tail,
+                            "Shape can't change for different arrays");
+    return ndim_tail;
+  }
+
+  // Returns full shape as read to/written from I/O, with 1-padding.
+  std::vector<int> GetPaddedShape(IndexDomain domain) const;
+  std::vector<int> GetPaddedShapeReversed(IndexDomain domain) const;
+  // nx accessors
+  std::vector<int> GetRawShape() const;
+  int GetDim(int i) const;
 
   VarInfo() = delete;
 
   // TODO(JMM): Separate this into an implementation file again?
   VarInfo(const std::string &label, const std::vector<std::string> &component_labels_,
-          int num_components, int nx6, int nx5, int nx4, int nx3, int nx2, int nx1,
-          Metadata metadata, bool is_sparse, bool is_vector)
-      : label(label), num_components(num_components), nx6(nx6), nx5(nx5), nx4(nx4),
-        nx3(nx3), nx2(nx2), nx1(nx1), tensor_rank(metadata.Shape().size()),
-        where(metadata.Where()), is_sparse(is_sparse), is_vector(is_vector) {
+          int num_components, std::array<int, VNDIM> nx, Metadata metadata,
+          const std::vector<TopologicalElement> topological_elements, bool is_sparse,
+          bool is_vector, const IndexShape &cellbounds)
+      : label(label), num_components(num_components), nx_(nx),
+        tensor_rank(metadata.Shape().size()), where(metadata.Where()),
+        topological_elements(topological_elements), is_sparse(is_sparse),
+        is_vector(is_vector), cellbounds(cellbounds), rnx_(nx_.rbegin(), nx_.rend()),
+        ntop_elems(topological_elements.size()), element_matters(ntop_elems > 1) {
     if (num_components <= 0) {
       std::stringstream msg;
       msg << "### ERROR: Got variable " << label << " with " << num_components
@@ -96,11 +163,22 @@ struct VarInfo {
     }
   }
 
-  explicit VarInfo(const std::shared_ptr<Variable<Real>> &var)
+  explicit VarInfo(const std::shared_ptr<Variable<Real>> &var,
+                   const IndexShape &cellbounds)
       : VarInfo(var->label(), var->metadata().getComponentLabels(), var->NumComponents(),
-                var->GetDim(6), var->GetDim(5), var->GetDim(4), var->GetDim(3),
-                var->GetDim(2), var->GetDim(1), var->metadata(), var->IsSparse(),
-                var->IsSet(Metadata::Vector)) {}
+                var->GetDim(), var->metadata(), var->GetTopologicalElements(),
+                var->IsSparse(), var->IsSet(Metadata::Vector), cellbounds) {}
+
+  static std::vector<VarInfo> GetAll(const VariableVector<Real> &vars,
+                                     const IndexShape &cellbounds);
+
+  bool operator==(const std::string &other) const { return other == label; }
+
+ private:
+  // TODO(JMM): Probably nx_ and rnx_ both not necessary... but it was
+  // easiest for me to reason about it this way.
+  std::array<int, VNDIM> nx_;
+  std::vector<int> rnx_;
 };
 
 struct SwarmVarInfo {
@@ -215,32 +293,35 @@ std::vector<T> FlattenBlockInfo(Mesh *pm, int shape, Function_t f) {
 }
 
 // mirror must be provided because copying done externally
-template <typename Data_t, typename idx_t, typename Function_t>
-void PackOrUnpackVar(MeshBlock *pmb, Variable<Real> *pvar, bool do_ghosts, idx_t &idx,
-                     std::vector<Data_t> &data, Function_t f) {
-  const auto &Nt = pvar->GetDim(6);
-  const auto &Nu = pvar->GetDim(5);
-  const auto &Nv = pvar->GetDim(4);
+template <typename idx_t, typename Function_t>
+void PackOrUnpackVar(const VarInfo &info, bool do_ghosts, idx_t &idx, Function_t f) {
   const IndexDomain domain = (do_ghosts ? IndexDomain::entire : IndexDomain::interior);
-  IndexRange kb, jb, ib;
-  if (pvar->metadata().Where() == MetadataFlag(Metadata::Cell)) {
-    kb = pmb->cellbounds.GetBoundsK(domain);
-    jb = pmb->cellbounds.GetBoundsJ(domain);
-    ib = pmb->cellbounds.GetBoundsI(domain);
-    // TODO(JMM): Add topological elements here
-  } else { // metadata none
-    kb = {0, pvar->GetDim(3) - 1};
-    jb = {0, pvar->GetDim(2) - 1};
-    ib = {0, pvar->GetDim(1) - 1};
+  // shape as written to or read from. contains additional padding
+  // in orthogonal directions.
+  // e.g., Face1-centered var is shape (N1+1)x(N2+1)x(N3+1)
+  // format is
+  // topological_elems x tensor_elems x block_elems
+  const auto shape = info.GetPaddedShapeReversed(domain);
+  // TODO(JMM): Should I hide this inside VarInfo?
+  auto [kb, jb, ib] = info.GetPaddedBoundsKJI(domain);
+  if (info.where == MetadataFlag({Metadata::None})) {
+    kb.s = 0;
+    kb.e = shape[4];
+    jb.s = 0;
+    jb.e = shape[5];
+    ib.s = 0;
+    ib.e = shape[6];
   }
-  for (int t = 0; t < Nt; ++t) {
-    for (int u = 0; u < Nu; ++u) {
-      for (int v = 0; v < Nv; ++v) {
-        for (int k = kb.s; k <= kb.e; ++k) {
-          for (int j = jb.s; j <= jb.e; ++j) {
-            for (int i = ib.s; i <= ib.e; ++i) {
-              f(idx, t, u, v, k, j, i);
-              idx++;
+  for (int topo = 0; topo < shape[0]; ++topo) {
+    for (int t = 0; t < shape[1]; ++t) {
+      for (int u = 0; u < shape[2]; ++u) {
+        for (int v = 0; v < shape[3]; ++v) {
+          for (int k = kb.s; k <= kb.e; ++k) {
+            for (int j = jb.s; j <= jb.e; ++j) {
+              for (int i = ib.s; i <= ib.e; ++i) {
+                f(idx, topo, t, u, v, k, j, i);
+                idx++;
+              }
             }
           }
         }